moving class content meta to individual file
