+++ subgraph/src/utils/index.ts
@@ -3,3 +3,4 @@ export * from "./entity/get";
 export * from "./bigNumbers";
 export * from "./constants";
 export * from "./entityString";
+export * from "./contentMeta";
